Fix import required by downstream commit.
