+++ lms/djangoapps/grades/tests/utils.py
@@ -3,6 +3,8 @@ Utilities for grades related tests
 """
 from contextlib import contextmanager
 from mock import patch
+from courseware.module_render import get_module
+from courseware.model_data import FieldDataCache
 from xmodule.graders import ProblemScore
 
 
@@ -24,3 +26,31 @@ def mock_get_score(earned=0, possible=1):
     with patch('lms.djangoapps.grades.new.subsection_grade.get_score') as mock_score:
         mock_score.return_value = ProblemScore(earned, possible, earned, possible, 1, True, None, None)
         yield mock_score
+
+
+def answer_problem(course, request, problem, score=1, max_value=1):
+    """
+    Records a correct answer for the given problem.
+
+    Arguments:
+        course (Course): Course object, the course the required problem is in
+        request (Request): request Object
+        problem (xblock): xblock object, the problem to be answered
+    """
+
+    user = request.user
+    grade_dict = {'value': score, 'max_value': max_value, 'user_id': user.id}
+    field_data_cache = FieldDataCache.cache_for_descriptor_descendents(
+        course.id,
+        user,
+        course,
+        depth=2
+    )
+    # pylint: disable=protected-access
+    module = get_module(
+        user,
+        request,
+        problem.scope_ids.usage_id,
+        field_data_cache,
+    )._xmodule
+    module.system.publish(problem, 'grade', grade_dict)
